+++ src/Buerokratt.Common.UnitTests/ServiceCollectionExtensionsTests.cs
@@ -1,5 +1,4 @@
 ﻿using Buerokratt.Common.CentOps;
-using Buerokratt.Common.Dmr.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using Xunit;
